src/requirements.txt
google-cloud-core
google-cloud
google-cloud-secret-manager